made audio record into separate component
