+++ playground/src/library/chat/chat-rendered.tsx
@@ -7,6 +7,7 @@ import {  Flex, Loader, Text, View, useTheme } from "@aws-amplify/ui-react";
 import { AgentChatMessage, AgentGraphQLBlock, AgentInnerDialogBlock, AgentJSONBlock, AgentPartialChatMessage, GraphQLResultBlock, UserChatMessage } from "./chat-items";
 import reactUseCookie from "react-use-cookie";
 import { useAgentConversationMetadata } from "../../apis/agent-api/hooks/useMetadata";
+import AudioRecorder from '../integrations/audio-recorder';
 
 
 function EnterUserSection () {
@@ -41,12 +42,9 @@ export function ChatRendered () {
     const chatInvokeQuery = useAgentApiInvokeQuery(chatId)
     setTimeout(() => Prism.highlightAll(), 100);
     useEffect(() => chatBottomRef.current?.scrollIntoView(), [events, conversationMetadata])
-    const [transcriptionText, setTranscriptionText] = useState('');
-    const [isRecording, setRecording] = useState(false);
+
     
-    function handleRecord() {
-        console.log("handle record")
-    }
+    
 
     if (agentObject.isUnloaded() || !agentObject.value || loadingConversation) {
         return <Loader/>
@@ -179,6 +177,11 @@ export function ChatRendered () {
         )
     }
 
+    //const handleRecordingComplete = (audioBlob) => {
+        // Handle the recorded audio blob here
+        // For example, sending it to the backend
+    //  };
+
     return (
         <View style={{height: 'calc(100vh - 230px)', overflowY: 'scroll'}}>
             
@@ -193,17 +196,8 @@ export function ChatRendered () {
                     <div ref={chatBottomRef}/>
                 </Flex>
             </View>
-            <View>
-                <label htmlFor="audioInputSelect">Choose a mic:</label>
-                <select name="" id="audioInputSelect"></select>
-
-                <button id="recordButton" onClick={handleRecord}>
-                    {isRecording ? 'Stop Recording' : 'Start Recording'}
-                </button>
-                <p>{isRecording ? 'Recording...' : ''}</p>
-                <button id="playButton" disabled>Play Recording</button>
-                <p>{transcriptionText}</p>
-            </View>
+            {/* <AudioRecorder onRecordingComplete={handleRecordingComplete} /> */}
+            <AudioRecorder />
         </View>
 
     )
